python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastembed>=0.7.3",
    "fastmcp>=2.13.0.2",
    "nats-py>=2.11.0",
    "numpyencoder>=0.3.2",
    "pydantic>=2.12.3",
    "requests>=2.32.5",
]
